Clean up code
Removed debug log statements, put unused scripts into "Unused" folder,
etc.
