+++ Assets/Scripts/RotatingGuard.cs
@@ -29,8 +29,6 @@ public class RotatingGuard : GuardController
         //Visualize raycast in Scene play. Doesn't affect gameplay
         //Debug.DrawRay(viewPoint.transform.position,(GameObject.FindWithTag("Player").transform.position - viewPoint.transform.position), Color.white, 0.0f, true);
 
-        print(GetAction());
-
         switch (GetAction())
         {
             case "idle":
