+++ Assets/Scripts/Unused.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5b400a00f448134c932a8aec489bb39
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
